requirements.txt
torch
torchaudio
joblib
python-dotenv